+++ project/build.properties
@@ -1,8 +1,1 @@
-#Project properties
-project.organization=net.stbbs.yasushi
-project.name=sbt-appengine-plugin
-sbt.version=0.7.5.RC0
-project.version=2.2
-def.scala.version=2.7.7
-build.scala.versions=2.7.7
-project.initialize=false
+sbt.version=0.10.0
